+++ crates/e2e/macro/src/codegen.rs
@@ -28,7 +28,7 @@ static BUILD_ONCE: Once = Once::new();
 
 thread_local! {
     // We save a mapping of `contract_manifest_path` to the built `*.contract` files.
-    // This is necessary so that not each individual `#[ink_e2e::e2e_tests]` starts
+    // This is necessary so that not each individual `#[ink_e2e::test]` starts
     // rebuilding the main contract and possibly specified `additional_contracts` contracts.
     pub static ALREADY_BUILT_CONTRACTS: RefCell<HashMap<String, String>> = RefCell::new(HashMap::new());
 }
@@ -73,7 +73,6 @@ impl InkE2ETest {
         };
 
         let ws_url = &self.test.config.ws_url();
-        let node_log = &self.test.config.node_log();
 
         let mut additional_contracts: Vec<String> =
             self.test.config.additional_contracts();
@@ -144,7 +143,7 @@ impl InkE2ETest {
                     let mut client = ::ink_e2e::Client::<
                         ::ink_e2e::PolkadotConfig,
                         ink::env::DefaultEnvironment
-                    >::new(&#ws_url, &#node_log).await;
+                    >::new(&#ws_url).await;
 
                     let __ret = {
                         #block
@@ -176,7 +175,6 @@ fn build_contract(manifest_path: &str) -> String {
             "+stable",
             "contract",
             "build",
-            "--skip-linting",
             "--output-json",
             &format!("--manifest-path={}", manifest_path),
         ])