+++ examples/delegator/lib.rs
@@ -127,7 +127,7 @@ mod delegator {
     mod e2e_tests {
         type E2EResult<T> = std::result::Result<T, Box<dyn std::error::Error>>;
 
-        #[ink_e2e::e2e_test(
+        #[ink_e2e::test(
             additional_contracts = "accumulator/Cargo.toml adder/Cargo.toml subber/Cargo.toml"
         )]
         async fn e2e_delegator(mut client: ink_e2e::Client<C, E>) -> E2EResult<()> {